@@ -1,21 +1,18 @@
 {
   "name": "dependencies-cli",
-  "version": "1.0.3",
+  "version": "2.0.0-beta",
   "author": "Andrew Fawcett, Justin Kalloor, Holger Kache",
   "bugs": "https://github.com/afawcett/dependencies-cli/issues",
   "dependencies": {
     "@oclif/config": "^1.6.27",
     "@oclif/errors": "^1.0.9",
     "@salesforce/command": "^1.2.0",
-    "@salesforce/core": "^1.2.0",
+    "@salesforce/core": "^1.3.2",
     "babel-preset-latest": "^6.24.1",
     "child-process": "^1.0.2",
     "child_process": "^1.0.2",
     "cors": "^2.8.5",
     "crypto": "^1.0.1",
-    "d3": "^5.9.1",
-    "d3-graphviz": "^2.6.1",
-    "dagre-d3": "^0.6.3",
     "debug": "^3.1.0",
     "dotenv": "^6.2.0",
     "express": "^4.16.4",
@@ -34,19 +31,17 @@
     "@oclif/tslint": "^1.0.2",
     "@salesforce/dev-config": "^1.1.0",
     "@types/chai": "^4.1.2",
+    "@types/graphlib-dot": "^0.6.1",
     "@types/jsforce": "^1.8.13",
     "@types/mocha": "^2.2.48",
     "@types/node": "^9.4.7",
-    "@types/d3": "^5.7.1",
-    "@types/graphlib-dot": "^0.6.1",
-    "@types/dagre-d3": "^0.4.39",
     "chai": "^4.1.2",
     "globby": "^8.0.1",
     "mocha": "^5.0.4",
     "sinon": "^6.0.0",
     "ts-node": "5",
     "tslint": "^5.9.1",
-    "typescript": "^2.7.2"
+    "typescript": "^3.3.0"
   },
   "engines": {
     "node": ">=8.4.0"
