@@ -1,8 +1,7 @@
-import { Project } from "@salesforce/core";
-import { SfdxCommand, core, S } from "@salesforce/command";
-import process = require('child_process');
+import { SfdxProject } from "@salesforce/core";
+import { SfdxCommand, core } from "@salesforce/command";
+// import process = require('child_process');
 
-import { SubscriberPackageVersion, InstalledSubscriberPackage } from "src/lib/NodeDefs";
 import { isNull } from "util";
 
 export default class Version extends SfdxCommand {
@@ -23,20 +22,20 @@ export default class Version extends SfdxCommand {
         conn.version = '45.0';
 
 
-        const project = await Project.resolve();
+        const project = await SfdxProject.resolve();
         const projectJson = await project.resolveProjectConfig();
         var subscriberPackage = [];
 
         if (!isNull(projectJson.packageDirectories)) {
-            for (const application of projectJson.packageDirectories) {
+       /*     for (const application of project.resolveProjectConfig) {
                 if (!isNull(application.dependencies)) {
                     for (const dependency of application.dependencies) {
                         subscriberPackage = await this.sfdxPackageVersionList(dependency);
                     }
                 }
             }
+            */
         }
-
 
         console.log(subscriberPackage);
 
@@ -50,7 +49,7 @@ export default class Version extends SfdxCommand {
     private async getSubscriberPackageVersions(connection: core.Connection) {
 
         let queryString = 'SELECT Dependencies FROM SubscriberPackageVersion WHERE Id=\'04t1U000003XEqdQAG\'';
-        const results = (await connection.tooling.query<SubscriberPackageVersion>(queryString)).records;
+        const results = (await connection.tooling.query<any>(queryString)).records;
 
         var dependency04tIds = [];
 
@@ -73,6 +72,8 @@ export default class Version extends SfdxCommand {
     Return
         map<string, array<string>> of packageNames to list of available package versions 
     */
+
+        /*
     private async sfdxPackageVersionList(packageName: string) {
 
         let cmd = "sfdx force:package:version:list --json"
@@ -80,7 +81,7 @@ export default class Version extends SfdxCommand {
         console.log(results);
 
         var dependency04tIds = [];
-/*
+
         for (const elem of results) {
             if (!isNull(elem.SubscriberPackageVersion) && (!isNull(elem.SubscriberPackageVersion.ids))) {
                 for (const id of elem.SubscriberPackageVersion.ids) {
@@ -88,10 +89,12 @@ export default class Version extends SfdxCommand {
                 }
             }
         }
-*/
+
         return dependency04tIds;
     }
+*/
 
+/*
     private async sh(cmd) {
         return new Promise((resolve, reject) => {
           process.exec(cmd, (err, stdout, stderr) => {
@@ -105,4 +108,6 @@ export default class Version extends SfdxCommand {
           });
         });
       }
+
+      */
 }
